account_manager/pyproject.toml
[project]
name = "account-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "seleniumbase>=4.36.0",
    "supabase>=2.28.3",
]
